migstreams: rename package to migsets
